// frontend/src/api/models/handlerProductListResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerProductListResponseStatus } from './handlerProductListResponseStatus'

/**
 * Product list item with basic information
 */
export interface HandlerProductListResponse {
  barcode?: string
  category_id?: string
  code?: string
  created_at?: string
  id?: string
  name?: string
  purchase_price?: number
  selling_price?: number
  sort_order?: number
  status?: HandlerProductListResponseStatus
  unit?: string
}