// frontend/src/api/models/handlerCreateSalesOrderItemInput.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Order item for creation
 */
export interface HandlerCreateSalesOrderItemInput {
  /**
   * @minLength 1
   * @maxLength 50
   */
  product_code: string
  product_id: string
  /**
   * @minLength 1
   * @maxLength 200
   */
  product_name: string
  quantity: number
  remark?: string
  /**
   * @minLength 1
   * @maxLength 20
   */
  unit: string
  unit_price: number
}